feat: text with link
