+++ packages/commons/src/icons/AddIcon.tsx
@@ -1,14 +1,13 @@
-import React from 'react';
-import { IIconProps } from '@tidy-ui/types';
+import React, { HTMLAttributes } from 'react';
 
-export default (props: IIconProps): JSX.Element => (
+export default (props: HTMLAttributes<SVGElement>): JSX.Element => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
-    className={props.className}
     focusable={false}
     aria-hidden="true"
     viewBox="0 0 24 24"
     fill="currentColor"
+    {...props}
   >
     <path d="M19 13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z" />
   </svg>
