+++ packages/commons/tests/icons/Icons.spec.tsx
@@ -16,31 +16,29 @@ import {
   DangerousIcon,
   InfoIcon,
   LaunchIcon,
+  LinkIcon,
   RotatingCircleIcon,
   WarningIcon,
 } from '../../src';
 
 const icons = [
   <AddCircleIcon />,
-  <AddCircleIcon outlined />,
+  <AddCircleIcon isOutlined />,
   <AddIcon />,
-  <AddIcon outlined />,
   <CancelIcon />,
-  <CancelIcon outlined />,
+  <CancelIcon isOutlined />,
   <CheckCircleIcon />,
-  <CheckCircleIcon outlined />,
+  <CheckCircleIcon isOutlined />,
   <CloseIcon />,
-  <CloseIcon outlined />,
   <DangerousIcon />,
-  <DangerousIcon outlined />,
+  <DangerousIcon isOutlined />,
   <InfoIcon />,
-  <InfoIcon outlined />,
+  <InfoIcon isOutlined />,
   <LaunchIcon />,
-  <LaunchIcon outlined />,
+  <LinkIcon />,
   <RotatingCircleIcon />,
-  <RotatingCircleIcon outlined />,
   <WarningIcon />,
-  <WarningIcon outlined />,
+  <WarningIcon isOutlined />,
 ];
 
 describe('Icons', () => {
