+++ packages/icons/src/ExpandMoreIcon.tsx
@@ -1,15 +1,16 @@
-import React from 'react';
-import { IIconProps } from '@tidy-ui/types';
+import React, { HTMLAttributes } from 'react';
 
-export default (props: IIconProps): JSX.Element => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    className={props.className}
-    focusable={false}
-    aria-hidden="true"
-    viewBox="0 0 24 24"
-    fill="currentColor"
-  >
-    <path d="M16.59 8.59 12 13.17 7.41 8.59 6 10l6 6 6-6z" />
-  </svg>
-);
+export default (props: HTMLAttributes<SVGElement>): JSX.Element => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      focusable={false}
+      aria-hidden="true"
+      viewBox="0 0 24 24"
+      fill="currentColor"
+      {...props}
+    >
+      <path d="M16.59 8.59 12 13.17 7.41 8.59 6 10l6 6 6-6z" />
+    </svg>
+  );
+};
