+++ packages/icons/src/WarningIcon.tsx
@@ -1,19 +1,22 @@
 import React from 'react';
 import { IIconProps } from '@tidy-ui/types';
 
-export default (props: IIconProps): JSX.Element => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    className={props.className}
-    focusable={false}
-    aria-hidden="true"
-    viewBox="0 0 24 24"
-    fill="currentColor"
-  >
-    {props.outlined ? (
-      <path d="M12 5.99 19.53 19H4.47L12 5.99M12 2 1 21h22L12 2zm1 14h-2v2h2v-2zm0-6h-2v4h2v-4z" />
-    ) : (
-      <path d="M1 21h22L12 2 1 21zm12-3h-2v-2h2v2zm0-4h-2v-4h2v4z" />
-    )}
-  </svg>
-);
+export default (props: IIconProps): JSX.Element => {
+  const { isOutlined, ...rest } = props;
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      focusable={false}
+      aria-hidden="true"
+      viewBox="0 0 24 24"
+      fill="currentColor"
+      {...rest}
+    >
+      {isOutlined ? (
+        <path d="M12 5.99 19.53 19H4.47L12 5.99M12 2 1 21h22L12 2zm1 14h-2v2h2v-2zm0-6h-2v4h2v-4z" />
+      ) : (
+        <path d="M1 21h22L12 2 1 21zm12-3h-2v-2h2v2zm0-4h-2v-4h2v4z" />
+      )}
+    </svg>
+  );
+};
