+++ packages/icons/tests/Icons.spec.tsx
@@ -23,6 +23,7 @@ import {
   InfoIcon,
   LaunchIcon,
   LightModeIcon,
+  LinkIcon,
   RemoveCircleIcon,
   RotatingCircleIcon,
   WarningIcon,
@@ -30,41 +31,36 @@ import {
 
 const icons = [
   <AddCircleIcon />,
-  <AddCircleIcon outlined />,
+  <AddCircleIcon isOutlined />,
   <AddIcon />,
-  <AddIcon outlined />,
   <CancelIcon />,
-  <CancelIcon outlined />,
+  <CancelIcon isOutlined />,
   <CheckCircleIcon />,
-  <CheckCircleIcon outlined />,
+  <CheckCircleIcon isOutlined />,
   <CloseIcon />,
-  <CloseIcon outlined />,
   <DangerousIcon />,
-  <DangerousIcon outlined />,
+  <DangerousIcon isOutlined />,
   <DarkModeIcon />,
-  <DarkModeIcon outlined />,
+  <DarkModeIcon isOutlined />,
   <EmailIcon />,
-  <EmailIcon outlined />,
+  <EmailIcon isOutlined />,
   <ExpandLessIcon />,
-  <ExpandLessIcon outlined />,
   <ExpandMoreIcon />,
-  <ExpandMoreIcon outlined />,
   <EyeIcon />,
-  <EyeIcon outlined />,
+  <EyeIcon isOutlined />,
   <GearIcon />,
-  <GearIcon outlined />,
+  <GearIcon isOutlined />,
   <InfoIcon />,
-  <InfoIcon outlined />,
+  <InfoIcon isOutlined />,
   <LaunchIcon />,
-  <LaunchIcon outlined />,
   <LightModeIcon />,
-  <LightModeIcon outlined />,
+  <LightModeIcon isOutlined />,
+  <LinkIcon />,
   <RemoveCircleIcon />,
-  <RemoveCircleIcon outlined />,
+  <RemoveCircleIcon isOutlined />,
   <RotatingCircleIcon />,
-  <RotatingCircleIcon outlined />,
   <WarningIcon />,
-  <WarningIcon outlined />,
+  <WarningIcon isOutlined />,
 ];
 
 describe('Icons', () => {
