+++ packages/text/src/Text.tsx
@@ -1,131 +1,6 @@
 import React, { forwardRef } from 'react';
-import { LinkIcon, useHash } from '@tidy-ui/commons';
-import { css, styled } from '@tidy-ui/theme';
+import { htmlElement, TextLink, TextLinkIcon, TextRoot } from './components';
 import { ITextProps } from './types';
-
-const Icon = styled(LinkIcon)`
-  display: none;
-  cursor: pointer;
-  height: 0.8em;
-  width: 0.8em;
-  transform: translate(0, 0.175em) rotate(-45deg);
-  opacity: 0.6;
-  margin-left: 0.2em;
-`;
-
-/**
- * Internal styled text
- *
- * @internal
- */
-const TextRoot = styled.div<ITextProps>`
-  ${({ theme: { palette, typography }, as, clr, bld, udl, itl, uc, lc, cc, dsb, exd, ctr, tnc }) => css`
-    display: block;
-    margin: 0;
-    font-size: ${typography[as!].fontSize};
-    ${!dsb &&
-    !clr &&
-    css`
-      color: ${palette.text.primary};
-    `}
-    ${!bld &&
-    css`
-      font-weight: ${typography[as!].fontWeight};
-    `}
-    letter-spacing: ${typography[as!].letterSpacing};
-    line-height: ${typography[as!].lineHeight};
-    ${!dsb &&
-    clr &&
-    css`
-      color: ${palette[clr][600]};
-    `}
-    ${bld &&
-    css`
-      font-weight: ${typography.fontWeightBold};
-    `}
-    ${udl &&
-    css`
-      text-decoration: underline;
-    `}
-    ${itl &&
-    css`
-      font-style: italic;
-    `}
-    ${uc &&
-    css`
-      text-transform: uppercase;
-    `}
-    ${lc &&
-    css`
-      text-transform: lowercase;
-    `}
-    ${cc &&
-    css`
-      text-transform: capitalize;
-    `}
-    ${dsb &&
-    css`
-      color: ${palette.text.disabled};
-    `}
-    ${exd &&
-    css`
-      width: 100%;
-    `}
-    ${ctr &&
-    css`
-      text-align: center;
-    `}
-    ${tnc &&
-    css`
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-    `}
-
-    &:hover ${Icon} {
-      display: inline-block;
-    }
-  `}
-`;
-
-/**
- * Internal Text root
- *
- * @internal
- */
-//const TextRoot = ({ children, ...rest }) => <StyledText {...rest}>{children}</StyledText>;
-
-const htmlElement = (props: ITextProps) => {
-  switch (props.as) {
-    case 'h1':
-    case 'hero':
-      return 'h1';
-    case 'h2':
-    case 'title1':
-      return 'h2';
-    case 'h3':
-    case 'title2':
-      return 'h3';
-    case 'h4':
-    case 'subtitle1':
-      return 'h4';
-    case 'h5':
-    case 'subtitle2':
-      return 'h5';
-    case 'h6':
-      return 'h6';
-    case 'p':
-    case 'body1':
-    case 'body2':
-      return 'p';
-    case 'span':
-      return 'span';
-    case 'code':
-      return 'code';
-    default:
-      return 'span';
-  }
-};
 
 /**
  * Text component has all the typography that you need on a page. USe this to represent
@@ -133,13 +8,17 @@ const htmlElement = (props: ITextProps) => {
  *
  */
 const Text = forwardRef<HTMLDivElement, ITextProps>((props, ref) => {
-  const { children, id, as, ...rest } = props;
+  const { children, as, href, ...rest } = props;
   const ele = htmlElement({ as });
-  const generatedId = useHash(children?.toString(), { maxLen: 8 });
+
   return (
-    <TextRoot role="text" ref={ref} as={ele} {...rest} id={id || generatedId}>
+    <TextRoot role="text" ref={ref} as={ele} {...rest}>
       {children}
-      {['h1', 'h2', 'h3', 'h4', 'h5', 'h6'].includes(ele) && <Icon />}
+      {href && (
+        <TextLink role="text-link" href={href}>
+          <TextLinkIcon />
+        </TextLink>
+      )}
     </TextRoot>
   );
 });
