+++ packages/text/stories/Text.stories.tsx
@@ -79,6 +79,9 @@ export const colorVariants = () => (
 
 export const others = () => (
   <Container>
+    <Text as="subtitle2" href="https://google.com">
+      Linked: The quick brown fox jumps over the lazy dog
+    </Text>
     <Text bld>bold: The quick brown fox jumps over the lazy dog</Text>
     <Text udl>underlined: The quick brown fox jumps over the lazy dog</Text>
     <Text itl>italicized: The quick brown fox jumps over the lazy dog</Text>
