+++ packages/text/tests/Text.spec.tsx
@@ -100,6 +100,9 @@ describe('Text', () => {
   it('All other forms of text', () => {
     const tree = render(
       <ThemeProvider theme={orchidLight}>
+        <Text as="subtitle2" href="https://google.com">
+          Linked: The quick brown fox jumps over the lazy dog
+        </Text>
         <Text bld>bold: The quick brown fox jumps over the lazy dog</Text>
         <Text udl>underlined: The quick brown fox jumps over the lazy dog</Text>
         <Text itl>italicized: The quick brown fox jumps over the lazy dog</Text>
