+++ packages/types/src/index.ts
@@ -5,7 +5,6 @@ export * from './theme/base';
 /**
  * Base theme contract
  *
- * @author Balu Praveen Datty <dev.badatt@gmail.com>
  * @public
  * @typedef {ITidyUIBaseTheme}
  */
@@ -17,7 +16,6 @@ export * from './theme/theme';
 /**
  * Theme contract
  *
- * @author Balu Praveen Datty <dev.badatt@gmail.com>
  * @public
  * @typedef {ITidyUITheme}
  */
@@ -26,7 +24,6 @@ export type ITidyUITheme = IITidyUITheme;
 /**
  * Color types for components
  *
- * @author Balu Praveen Datty <dev.badatt@gmail.com>
  * @public
  * @typedef {TColor}
  */
@@ -35,7 +32,6 @@ export type TColor = 'major' | 'minor' | 'neutral' | 'info' | 'success' | 'warni
 /**
  * Size types for components
  *
- * @author Balu Praveen Datty <dev.badatt@gmail.com>
  * @public
  * @typedef {TSize}
  */
@@ -44,7 +40,6 @@ export type TSize = 'xxs' | 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
 /**
  * Accent types for components
  *
- * @author Balu Praveen Datty <dev.badatt@gmail.com>
  * @public
  * @typedef {TAccent}
  */
@@ -53,12 +48,11 @@ export type TAccent = 'major' | 'minor' | 'neutral' | 'info' | 'success' | 'warn
 /**
  * Common props for Icon components
  *
- * @author Balu Praveen Datty <dev.badatt@gmail.com>
  * @public
  * @interface IIconProps
  * @typedef {IIconProps}
  * @augments {HTMLAttributes<HTMLDivElement>}
  */
-export interface IIconProps extends HTMLAttributes<HTMLDivElement> {
-  outlined?: boolean;
+export interface IIconProps extends HTMLAttributes<SVGElement> {
+  isOutlined?: boolean;
 }
